+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Swagger comments on by default
 - 409 no longer shown on swagegr comments for POST
 - No more save successful check on delete command
+- Seeders have a sub `DummyData` directory
 
 ### Fixed